refactor: Move functions to their own js file
