+++ src/deck.js
@@ -1,0 +1,17 @@
+const { prototypeData } = require('./data');
+const { createCard, evaluateGuess } = require('./card');
+
+function createDeck(arr) {
+  return {
+    cards: arr,
+  };
+}
+
+function countCards(arr) {
+  return arr.cards.length === 0 ? 'This deck is empty!' : arr.cards.length;
+}
+
+module.exports = {
+  createDeck,
+  countCards,
+};
