+++ src/game.js
@@ -1,9 +1,12 @@
 const data = require('./data');
 const prototypeQuestions = data.prototypeData;
 const util = require('./util');
+const { createDeck, countCards, createRound, createCard } = require('./card');
 
 function printMessage(deck) {
-  console.log(`Welcome to FlashCards! You are playing with ${countCards(deck)} cards.
+  console.log(`Welcome to FlashCards! You are playing with ${countCards(
+    deck
+  )} cards.
   -----------------------------------------------------------------------`);
 }
 
@@ -11,4 +14,15 @@ function printQuestion(round) {
   util.main(round);
 }
 
-module.exports = { printMessage, printQuestion };
+function start() {
+  const deck = createDeck(data.prototypeData);
+  const round = createRound(deck);
+  printMessage(deck);
+  printQuestion(round);
+}
+
+module.exports = {
+  printMessage,
+  printQuestion,
+  start,
+};
